Remove atomics, use simple operations

+++ src/xmidi32_driver.h
@@ -4,7 +4,6 @@
 #include "xmidi32_config.h"
 #include "xmidi32_types.h"
 #include "xmidi32_backend.h"
-#include "xmidi32_critical.h"
 
 typedef int32_t HTIMER;
 typedef int32_t HDRIVER;
@@ -68,7 +67,7 @@ extern struct ctrl_log global_controls;
 extern uint8_t global_program[NUM_CHANS];
 extern uint8_t global_pitch_l[NUM_CHANS];
 extern uint8_t global_pitch_h[NUM_CHANS];
-extern uint8_t active_notes[NUM_CHANS];
+extern uint32_t active_notes[NUM_CHANS];
 extern uint8_t lock_status[NUM_CHANS];
 extern uint16_t init_OK;
 extern uint32_t trigger_active;
